# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.pagedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: Backend API implementing ISO/IEC 19770-10 Grade A++ RBAC model
  title: IT Asset & Compliance API
  version: "1.1"
paths:
...
  /services/availability:
    get:
      parameters:
      - description: Filter by service code
        in: query
        name: service_code
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.pagedResponse'
      security:
      - BearerAuth: []
      summary: List data availability layanan
      tags:
      - Vendor & Service
...
  /vendors/performance:
    get:
      parameters:
      - description: Filter by vendor name
        in: query
        name: vendor
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.pagedResponse'
      security:
      - BearerAuth: []
      summary: List performa vendor
      tags:
      - Vendor & Service
swagger: "2.0"